Create README FOR INFO
Hello!! I want to tell you something about me in this file. I've started the lockdown with a goal in my mind, I wanted to learn something new and I think that not by accident I've discovered Python and Pygame. This is the first project that I've done. I am proud of this? To be honest not really, it is my first attempt to make a game and it could be always better, BUT I am proud of me, proud of the late hours spent in PyCharm and proud that I've managed to learn something new and fun. All this process was an adventure, from the day that I've found an idea, to the days and night tryin' to figure out all those annoying errors and finally this day when I try to find in me the courage to share this with you. If you've managed to read all this boring personal stuff I want you to know that in those times of change and uncertainty we must try to change, to evolve. I wish you luck and I am sure that you can achieve your goals. THANK YOU!!